refactor: use int64 timestamps (#189)
* refactor: use int64 timestamps
* fix: changed PB Timestamp index to 10

+++ examples/basic2/main.go
@@ -84,7 +84,7 @@ func write(ctx context.Context, wakuNode *node.WakuNode, msgContent string) {
 	contentTopic := protocol.NewContentTopic("basic2", 1, "test", "proto")
 
 	var version uint32 = 0
-	var timestamp float64 = utils.GetUnixEpoch()
+	var timestamp int64 = utils.GetUnixEpoch()
 
 	p := new(node.Payload)
 	p.Data = []byte(wakuNode.ID() + ": " + msgContent)
